Import java-nachos sources originally by Peter Druschel, Rice University. The files in this version are mostly dated August 1, 1999, and I got them from the following URL:
	http://saathi.ncst.ernet.in/~pgdst/coos/nachos/started.html

I am not able to find a more direct reference to Druschel's sources at this
time.

The Berkeley cs162 home page at:

	http://www-inst.EECS.Berkeley.EDU/~cs162/

has a different version of Nachos in Java.  It has some advantages,
namely that it seems to be more idiomatic for Java and has Javadoc
documentation, but they have also made a fair number of changes since
the original version of Nachos that I liked.  Some of these changes
seem to complicate things, rather than maintain the simplicity of the
original, so at this time I have decided to use the Druschel version.