Changes to make it possible to create MIPS/NOFF binaries.
The bin directory and coff2noff utility was missing in the distribution
tarball -- recover it from the old Berkeley distribution.
This requires a native C compiler to build, so I plan to rewrite it
in Java at some point.

The syscall.h file was also missing.  It is needed for compiling start.s.
Recover this from the old Berkeley distribution, too, and edit it to
add the Remove system call.

Add include directory as a home for the syscall.h and copyright.h files
used by C/assembler code in bin and test.  Only the C programs in test
actually include syscall.h, but it is really part of the Nachos API,
not a test program, so it seems like the test directory is the wrong
place for it.  Userprog is probably also wrong, now that all is in Java.

The endianness assumed by the NoffHeader class did not match what was
produced by coff2noff on my Intel system.  Reverse the byte order.
This will be better once Coff2Noff is rewritten in Java.

Recompile all tests to get new binaries that account for the different
endianness.